# docs/swagger.yaml
basePath: /api
definitions:
  model.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  model.CartItemDetail:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      subtotal:
        type: number
    type: object
  model.CartResponse:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.CartItemDetail'
        type: array
      total:
        type: number
    type: object
  model.CreateOrderRequest:
    properties:
      shipping_address:
        type: string
    required:
    - shipping_address
    type: object
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.OrderItemDetail:
    properties:
      name:
        type: string
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      subtotal:
        type: number
    type: object
  model.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.OrderItemDetail'
        type: array
      shipping_address:
        type: string
      status:
        type: string
      total_amount:
        type: number
    type: object
  model.OrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/model.OrderResponse'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: ini_email@ordent.com
    name: API Support
  description: API for E-Commerce platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Commerce API
  version: "1.0"
paths:
...
  /cart:
    get:
      consumes:
      - application/json
      description: Get the current user's shopping cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CartResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Add a product to the user's shopping cart
      parameters:
      - description: Item to add
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - cart
  /cart/items/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an item from the user's shopping cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - cart
  /orders:
    get:
      consumes:
      - application/json
      description: Get a list of user's orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrdersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order from cart items
      parameters:
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"